+++ mainapp/1_IVR_Data_Cleaner_Pre_Processor_App.py
@@ -1,8 +1,8 @@
 import streamlit as st
 import pandas as pd
 from datetime import datetime
-from modules.security_utils import check_password
-from modules.data_cleaner_utils_page1 import process_file
+from modules_backend.security_utils import check_password
+from modules_backend.data_cleaner_utils_page1 import process_file
 from PIL import Image
 import numpy as np
